Reorganize is_hibernation_enabled_for_vm()
With this change, we first check /sys/power/disk for supported hibernation
modes, and print out the ones that we know will work to hibernate a machine,
but bail out if none of those are found.

Then, we probe /sys/bus/vmbus/hibernation to determine if hibernation has
been enabled in Hyper-V as well.  If so, then we proceed as usual.  If not,
we don't continue.

At this point, we also verify if /sys/bus/vmbus exists as a way to determine
if we're running on a Hyper-V VM.  This is just to print a better log message
to aid in debugging should the hibernation file not be present.